ai_service/requirements.txt
git+https://@github.com/fhuadeen/loh_ai.git
# git+https://<github-access-key>@github.com/fhuadeen/loh_utils.git
Flask==3.0.3
Flask-RESTful==0.3.10
python-dotenv==1.0.0
flask-swagger-ui==4.11.1
flasgger==0.9.7.1